+++ src/index.js
@@ -19,7 +19,7 @@ import "./Components/Tree/kiwi-treelist-item"
 import { injectStyle } from "./Utility/StyleInjection"
 import { showToast } from "./Utility/Notifications"
 import { enableGlobalTooltips } from "./Utility/Tooltip"
-import { openWindow, confirm, showSpinner } from "./Utility/WindowManager"
+import { openWindow, confirm, showSpinner, alert, prompt } from "./Utility/WindowManager"
 
 //Export
-export { injectStyle, showToast, enableGlobalTooltips, openWindow, confirm, showSpinner }
+export { injectStyle, showToast, enableGlobalTooltips, openWindow, confirm, showSpinner, alert, prompt }